Modify Transaction component
